+++ src/index.ts
@@ -1,5 +1,2 @@
-import loadScript = require("./load-script");
-import loadStylesheet = require("./load-stylesheet");
-
-export { loadScript };
-export { loadStylesheet };
+export { default as loadScript } from "./lib/load-script";
+export { default as loadStylesheet } from "./lib/load-stylesheet";