+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+- BREAKING CHANGES
+  - Update exports to support ES6 syntax
+  - Support both `import` and `require` syntax
+  - Migrate from `tsc` to vite for compiling & build
+  - Migrate tests from Jest to Vitest
+
 # 1.0.1
 
 - Dependabot Updates
@@ -5,6 +13,7 @@
 # 1.0.0-beta.1
 
 - BREAKING CHANGES
+
   - Update Node to v18
 
 - DevDependency Updates